// u-client/src/app/features/inventory/model/Report/productCustomerYearlyReport.ts
import { MyBarChart } from './barChart';
import { CustomerMonthlyReport } from './customerMonthlyReport';

export class ProductCustomerYearlyReport {
  prodid: number;
  prodname: number;
  cid: number;
  name: string;
  saleqty: number;
  saleamt: number;
  avgSalePrice: number;
  profit: number;

  rupSaleAmt :string;
  rupAvgSalePrice: string;
  rupProfit: string;




}