// u-client/src/app/features/inventory/model/Report/productSupplierYearlyReport.ts
import { MyBarChart } from './barChart';
import { CustomerMonthlyReport } from './customerMonthlyReport';

export class ProductSupplierYearlyReport {
  prodid: number;
  prodname: number;
  cid: number;
  name: string;
  
  purchaseqty: number;
  purchaseamt: number;
  avgCostPrice: number;

  rupPurchaseAmt :string;
  rupAvgCostPrice: string;

}